fix: Guid inputs validation

+++ server/src/Domain/Handlers/ItemHandler.cs
@@ -38,18 +38,17 @@ public class ItemHandler :
     public GenericCommandResult Handle(MarkItemAsPurchasedCommand command)
     {
         var item = _repository.GetAndValidateOwner(command.Id, command.UserId).Result;
+        if (item == null) return new GenericCommandResult(false, "Item not found", item, null);
         item.IsPurchased = true;
         item.PurchasedAt = DateTime.Now;
-
         _repository.Update(item);
-        
-        return new GenericCommandResult(true, "Item was purchased", item,null);
+        return new GenericCommandResult(true, "Item was purchased", item, null);
     }
 
     public GenericCommandResult Handle(DeleteItemCommand command)
     {
         var item = _repository.GetAndValidateOwner(command.ItemId, command.UserId).Result;
-
+        if (item == null) return new GenericCommandResult(false, "Item not found", item, null);
         _repository.Delete(item.Id);
         return new GenericCommandResult(true, "Item was deleted", null,null);
     }
